+++ lib/fcrm_chat_sdk.dart
@@ -9,6 +9,7 @@ export 'src/config/chat_config.dart';
 
 // Models
 export 'src/models/message.dart';
+export 'src/models/message_rating.dart';
 export 'src/models/chat_app_config.dart';
 export 'src/models/browser.dart';
 export 'src/models/paginated_messages.dart';